requirements.txt
fastapi
uvicorn
sqlalchemy
python-telegram-bot
requests
beautifulsoup4
selenium
webdriver-manager
schedule
python-dotenv
pydantic
pydantic-settings
python-multipart
aiofiles
fuzzywuzzy[speedup]
httpx
